requirements.txt
python-rtmidi
numpy
pyaudio #on mac requires: brew install portaudio
aubio

